refactor(reporting): Split MCP server into core + daemon; drop Halo prefix
Split the single file into a non-flags core and a flags daemon, matching
the PopulationRequisitionFulfiller pattern:
  - ReportingMcpServer.kt: the ReportingMcpServer factory object,
    installReportingMcp, and bearerToken (no picocli).
  - ReportingMcpServerDaemon.kt: ReportingMcpServerDaemon (Runnable),
    McpServerFlags, and main().
Drop the Halo prefix across file names, BUILD library/binary/image
targets, and the main_class. Rename McpServerTest to
ReportingMcpServerTest.

Issue: #3834
