+++ src/main/kotlin/org/wfanet/measurement/reporting/mcp/BUILD.bazel
@@ -3,8 +3,8 @@ load("@wfa_rules_kotlin_jvm//kotlin:defs.bzl", "kt_jvm_library")
 load("//src/main/docker:macros.bzl", "java_image")
 
 kt_jvm_library(
-    name = "halo_reporting_mcp_server",
-    srcs = ["HaloReportingMcpServer.kt"],
+    name = "reporting_mcp_server",
+    srcs = ["ReportingMcpServer.kt"],
     visibility = [
         "//src/test/kotlin/org/wfanet/measurement/reporting/mcp:__subpackages__",
     ],
@@ -15,7 +15,6 @@ kt_jvm_library(
         "//imports/java/io/ktor/server:sse",
     ],
     deps = [
-        "//imports/java/io/ktor/server:cio",
         "//imports/java/io/ktor/server:cors",
         "//imports/java/io/modelcontextprotocol/kotlin/sdk:server",
         "//src/main/kotlin/org/wfanet/measurement/reporting/mcp/grpc:reporting_public_api_client",
@@ -23,6 +22,16 @@ kt_jvm_library(
         "//src/main/kotlin/org/wfanet/measurement/reporting/mcp/tools:event_group_tools",
         "//src/main/kotlin/org/wfanet/measurement/reporting/mcp/tools:iqf_tools",
         "//src/main/kotlin/org/wfanet/measurement/reporting/mcp/tools:reporting_set_tools",
+    ],
+)
+
+kt_jvm_library(
+    name = "reporting_mcp_server_daemon",
+    srcs = ["ReportingMcpServerDaemon.kt"],
+    deps = [
+        ":reporting_mcp_server",
+        "//imports/java/io/ktor/server:cio",
+        "//src/main/kotlin/org/wfanet/measurement/reporting/mcp/grpc:reporting_public_api_client",
         "//src/main/proto/wfa/measurement/reporting/v2alpha:basic_reports_service_kt_jvm_grpc_proto",
         "//src/main/proto/wfa/measurement/reporting/v2alpha:event_groups_service_kt_jvm_grpc_proto",
         "//src/main/proto/wfa/measurement/reporting/v2alpha:impression_qualification_filters_service_kt_jvm_grpc_proto",
@@ -35,14 +44,14 @@ kt_jvm_library(
 )
 
 java_binary(
-    name = "HaloReportingMcpServer",
-    main_class = "org.wfanet.measurement.reporting.mcp.HaloReportingMcpServerKt",
-    runtime_deps = [":halo_reporting_mcp_server"],
+    name = "ReportingMcpServerDaemon",
+    main_class = "org.wfanet.measurement.reporting.mcp.ReportingMcpServerDaemonKt",
+    runtime_deps = [":reporting_mcp_server_daemon"],
 )
 
 java_image(
-    name = "halo_reporting_mcp_server_image",
-    binary = ":HaloReportingMcpServer",
-    main_class = "org.wfanet.measurement.reporting.mcp.HaloReportingMcpServerKt",
+    name = "reporting_mcp_server_daemon_image",
+    binary = ":ReportingMcpServerDaemon",
+    main_class = "org.wfanet.measurement.reporting.mcp.ReportingMcpServerDaemonKt",
     visibility = ["//src:docker_image_deployment"],
 )
