+++ src/main/kotlin/org/wfanet/measurement/reporting/mcp/ReportingMcpServerDaemon.kt
@@ -22,22 +22,8 @@ import io.grpc.ClientCall
 import io.grpc.ClientInterceptor
 import io.grpc.ClientInterceptors
 import io.grpc.MethodDescriptor
-import io.ktor.http.HttpHeaders
-import io.ktor.http.HttpMethod
-import io.ktor.http.HttpStatusCode
-import io.ktor.server.application.Application
-import io.ktor.server.application.install
 import io.ktor.server.cio.CIO
 import io.ktor.server.engine.embeddedServer
-import io.ktor.server.plugins.cors.routing.CORS
-import io.ktor.server.response.respondText
-import io.ktor.server.routing.get
-import io.ktor.server.routing.routing
-import io.modelcontextprotocol.kotlin.sdk.server.Server
-import io.modelcontextprotocol.kotlin.sdk.server.ServerOptions
-import io.modelcontextprotocol.kotlin.sdk.server.mcpStatelessStreamableHttp
-import io.modelcontextprotocol.kotlin.sdk.types.Implementation
-import io.modelcontextprotocol.kotlin.sdk.types.ServerCapabilities
 import java.time.Duration
 import java.util.concurrent.TimeUnit
 import kotlin.properties.Delegates
@@ -47,117 +33,20 @@ import org.wfanet.measurement.common.grpc.TlsFlags
 import org.wfanet.measurement.common.grpc.buildMutualTlsChannel
 import org.wfanet.measurement.common.grpc.withVerboseLogging
 import org.wfanet.measurement.reporting.mcp.grpc.ReportingPublicApiClient
-import org.wfanet.measurement.reporting.mcp.tools.registerBasicReportTools
-import org.wfanet.measurement.reporting.mcp.tools.registerEventGroupTools
-import org.wfanet.measurement.reporting.mcp.tools.registerIqfTools
-import org.wfanet.measurement.reporting.mcp.tools.registerReportingSetTools
 import org.wfanet.measurement.reporting.v2alpha.BasicReportsGrpcKt.BasicReportsCoroutineStub
 import org.wfanet.measurement.reporting.v2alpha.EventGroupsGrpcKt.EventGroupsCoroutineStub
 import org.wfanet.measurement.reporting.v2alpha.ImpressionQualificationFiltersGrpcKt.ImpressionQualificationFiltersCoroutineStub
 import org.wfanet.measurement.reporting.v2alpha.ReportingSetsGrpcKt.ReportingSetsCoroutineStub
 import picocli.CommandLine
 
-private const val SERVER_NAME = "HaloReportingMcpServer"
-private const val SERVER_VERSION = "0.1.0"
-private const val BEARER_PREFIX = "Bearer "
-
-/**
- * Configures this [Application] as the Halo Reporting MCP server.
- *
- * Stateless Streamable HTTP: the SDK helper creates a fresh [Server] per POST (no session state),
- * so any replica can serve any request and pod restarts lose nothing. The helper installs
- * ContentNegotiation(McpJson) and SSE itself.
- *
- * DNS rebinding protection is enabled only when [allowedHosts] is non-empty (e.g. the in-cluster
- * service hostnames); otherwise the Host header is not checked.
- *
- * The bearer token is read per request and resolved lazily inside each tool call: a missing token
- * throws [IllegalArgumentException], which the tool error handler turns into a clean tool error
- * rather than a 500.
- */
-fun Application.installReportingMcp(
-  apiClient: ReportingPublicApiClient,
-  allowedHosts: List<String> = emptyList(),
-) {
-  install(CORS) {
-    anyHost()
-    allowNonSimpleContentTypes = true
-    allowMethod(HttpMethod.Options)
-    allowMethod(HttpMethod.Post)
-    allowHeader(HttpHeaders.ContentType)
-    allowHeader(HttpHeaders.Authorization)
-    allowHeader("Mcp-Protocol-Version")
-    exposeHeader("Mcp-Protocol-Version")
-  }
-
-  mcpStatelessStreamableHttp(
-    enableDnsRebindingProtection = allowedHosts.isNotEmpty(),
-    allowedHosts = allowedHosts.ifEmpty { null },
-  ) {
-    val authorizationHeader = call.request.headers[HttpHeaders.Authorization]
-    ReportingMcpServer.createServer(apiClient) { bearerToken(authorizationHeader) }
-  }
-
-  routing { get("/healthz") { call.respondText("OK", status = HttpStatusCode.OK) } }
-}
-
-/**
- * Extracts the bearer token from an `Authorization` header value, forwarded to the Reporting API.
- *
- * @throws IllegalArgumentException if the header is missing or has no non-blank bearer token
- */
-private fun bearerToken(authorizationHeader: String?): String {
-  val token =
-    authorizationHeader
-      ?.takeIf { it.startsWith(BEARER_PREFIX, ignoreCase = true) }
-      ?.substring(BEARER_PREFIX.length)
-      ?.trim()
-  return requireNotNull(token?.takeUnless(String::isEmpty)) {
-    "Missing bearer token in Authorization header"
-  }
-}
-
-/**
- * Factory for the Reporting MCP [Server].
- *
- * Builds the server from already-constructed dependencies and has no knowledge of flags. The
- * flags-based entry point is [ReportingMcpServerFromFlags].
- */
-object ReportingMcpServer {
-  /** Builds the MCP [Server] with all Reporting tools registered. */
-  fun createServer(apiClient: ReportingPublicApiClient, getBearerToken: () -> String): Server {
-    val server =
-      Server(
-        serverInfo = Implementation(name = SERVER_NAME, version = SERVER_VERSION),
-        options =
-          ServerOptions(
-            capabilities =
-              ServerCapabilities(
-                // Stateless mode: no server-to-client push (logging/notifications), so
-                // only the tools capability is advertised.
-                tools = ServerCapabilities.Tools(listChanged = false)
-                // TODO(#3834): Add prompts capability in follow-up PR.
-              )
-          ),
-      )
-
-    server.registerBasicReportTools(apiClient, getBearerToken)
-    server.registerEventGroupTools(apiClient, getBearerToken)
-    server.registerReportingSetTools(apiClient, getBearerToken)
-    server.registerIqfTools(apiClient, getBearerToken)
-
-    return server
-  }
-}
-
 /** Builds the runtime dependencies from command-line flags and starts the MCP server. */
 @CommandLine.Command(
-  name = SERVER_NAME,
-  description = ["MCP server for the Halo Reporting v2alpha public API."],
+  name = "ReportingMcpServerDaemon",
+  description = ["MCP server for the Reporting v2alpha public API."],
   mixinStandardHelpOptions = true,
   showDefaultValues = true,
 )
-class ReportingMcpServerFromFlags : Runnable {
+class ReportingMcpServerDaemon : Runnable {
   @CommandLine.Mixin private lateinit var tlsFlags: TlsFlags
   @CommandLine.Mixin private lateinit var mcpServerFlags: McpServerFlags
 
@@ -271,4 +160,4 @@ class McpServerFlags {
   var debugVerboseGrpcClientLogging by Delegates.notNull<Boolean>()
 }
 
-fun main(args: Array<String>) = commandLineMain(ReportingMcpServerFromFlags(), args)
+fun main(args: Array<String>) = commandLineMain(ReportingMcpServerDaemon(), args)
