+++ src/test/kotlin/org/wfanet/measurement/reporting/mcp/BUILD.bazel
@@ -1,15 +1,15 @@
 load("@wfa_rules_kotlin_jvm//kotlin:defs.bzl", "kt_jvm_test")
 
 kt_jvm_test(
-    name = "McpServerTest",
-    srcs = ["McpServerTest.kt"],
-    test_class = "org.wfanet.measurement.reporting.mcp.McpServerTest",
+    name = "ReportingMcpServerTest",
+    srcs = ["ReportingMcpServerTest.kt"],
+    test_class = "org.wfanet.measurement.reporting.mcp.ReportingMcpServerTest",
     deps = [
         "//imports/java/io/ktor/server:cio",
         "//imports/java/io/modelcontextprotocol/kotlin/sdk:client",
         "//imports/java/io/modelcontextprotocol/kotlin/sdk:server",
         "//imports/java/io/modelcontextprotocol/kotlin/sdk:testing",
-        "//src/main/kotlin/org/wfanet/measurement/reporting/mcp:halo_reporting_mcp_server",
+        "//src/main/kotlin/org/wfanet/measurement/reporting/mcp:reporting_mcp_server",
         "//src/main/kotlin/org/wfanet/measurement/reporting/mcp/grpc:reporting_public_api_client",
         "//src/main/proto/wfa/measurement/reporting/v2alpha:basic_report_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/reporting/v2alpha:basic_reports_service_kt_jvm_grpc_proto",
