+++ src/test/kotlin/org/wfanet/measurement/reporting/mcp/tools/BUILD.bazel
@@ -6,7 +6,7 @@ kt_jvm_test(
     test_class = "org.wfanet.measurement.reporting.mcp.tools.ToolsIntegrationTest",
     deps = [
         "//imports/java/io/modelcontextprotocol/kotlin/sdk:server",
-        "//src/main/kotlin/org/wfanet/measurement/reporting/mcp:halo_reporting_mcp_server",
+        "//src/main/kotlin/org/wfanet/measurement/reporting/mcp:reporting_mcp_server",
         "//src/main/kotlin/org/wfanet/measurement/reporting/mcp/grpc:reporting_public_api_client",
         "//src/main/proto/wfa/measurement/reporting/v2alpha:basic_report_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/reporting/v2alpha:basic_reports_service_kt_jvm_grpc_proto",
